+++ core/entity/components/states/BaseState.gd
@@ -1,0 +1,12 @@
+extends Node
+
+var entity: Node
+
+func enter() -> void:
+	pass
+
+func exit() -> void:
+	pass
+
+func update(_delta: float) -> void:
+	pass